GCS/Uploader emit signal with new connected boards data
